@@ -7,7 +7,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "prisma": "prisma"
+    "prisma": "prisma",
+    "postinstall": "prisma generate"
   },
   "dependencies": {
     "@auth/mongodb-adapter": "^3.1.0",
@@ -15,16 +16,27 @@
     "@capacitor/core": "^6.1.0",
     "@capacitor/ios": "^6.1.0",
     "@hookform/resolvers": "^3.4.2",
+    "@libsql/client": "^0.6.2",
+    "@prisma/adapter-libsql": "^5.17.0",
     "@prisma/client": "^5.17.0",
+    "@types/lodash.debounce": "^4.0.9",
     "autoprefixer": "^10.4.19",
     "axios": "^1.7.2",
     "babel-preset-es2015": "^6.24.1",
     "bcrypt": "^5.1.1",
     "canvas-confetti": "^1.9.3",
     "cloudinary": "^2.2.0",
+    "csv-parser": "^3.0.0",
+    "csv-writer": "^1.6.0",
     "date-fns": "^3.6.0",
+    "debounce": "^2.1.0",
     "dotenv": "^16.4.5",
+    "drizzle-orm": "^0.32.0",
+    "ioredis": "^5.4.1",
+    "json2csv": "^6.0.0-alpha.2",
     "keen-slider": "^6.8.6",
+    "libsql-stateless-easy": "^1.7.0",
+    "lodash.debounce": "^4.0.8",
     "lru-cache": "^11.0.0",
     "mongodb": "^6.6.2",
     "mongodb-data-api": "^0.4.0",
@@ -33,17 +45,20 @@
     "next-auth": "^4.24.7",
     "next-cloudinary": "^6.6.0",
     "node-cache": "^5.1.2",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.51.5",
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.2.1",
     "react-select": "^5.8.0",
     "redis": "^4.6.15",
+    "sqlite3": "^5.1.7",
+    "uuid": "^10.0.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
     "@capacitor/cli": "^6.1.0",
+    "drizzle-kit": "^0.23.0",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
     "html-loader": "^5.0.0",
