@@ -1,4 +1,3 @@
-'use client';
 import { useContext, useEffect, useState } from 'react';
 import { inputsContext } from './Context';
 import { useSession } from 'next-auth/react';
@@ -6,23 +5,30 @@ import { useSession } from 'next-auth/react';
 export default function CurrentUser() {
   const { profile_image } = useContext(inputsContext);
   const [user, setUser] = useState();
-  // const session = useSession();
-  // console.log('user', user);
+  const { data: session, status } = useSession();
+
   useEffect(() => {
-    getUserData();
-  }, [profile_image?.image]);
+    if (status === 'authenticated') {
+      getUserData();
+    }
+  }, [status, profile_image?.image]);
 
   async function getUserData() {
-    const response = await fetch('/api/user');
-    const json = await response?.json();
+    if (session) {
+      const email = session?.user?.email;
+      // console.log('email', email);
+      const response = await fetch(`/api/user?email=${email}`);
+      const json = await response?.json();
+      // console.log('json', json);
+      if (response.ok) {
+        if (typeof window !== 'undefined') {
+          localStorage.setItem('CurrentUser', JSON.stringify(json));
+        }
 
-    if (response.ok) {
-      if (typeof window !== 'undefined') {
-        localStorage.setItem('CurrentUser', JSON.stringify(json[0]));
+        setUser(json);
       }
-
-      setUser(json[0]);
     }
   }
+
   return { ...user };
 }
